🚀 chore(addon.xml): update version to 1.7.2 🔨 refactor(default.py): remove platform logs from main function and add them to a separate function The addon.xml file has been updated to version 1.7.2. This is a minor version update that includes bug fixes and performance improvements.
The platform logs have been removed from the main function and added to a separate function called platform_logs(). This improves the readability of the code and makes it easier to maintain. The platform_logs() function now logs additional information about the operating system by using the os.uname() function.